Initial commit.
Terminal will be a platformer game built completely with python3 and
curses.  The game will be based in a 80x24 window (for maximum compatibility
with older systems), and this window will be bordered with unicode box
symbols. This commit contains a demo of this bordered window with a display
of ascii art.

Additions:
    - Resources folder created. This will store all game resources, e.g.
      levels and ascii-art.
    - The library terminalDraw will handle all drawing functions, e.g. adding
      text, drawing levels, drawing sprites.
    - The library terminalWindowUtils will handle all window utilites, e.g.
      adding a border, adding menues, etc.